make a reusable axe pytest fixture
